requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
gprof2dot==2024.6.6
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
urllib3==2.3.0
